requirements.txt
matplotlib==3.3.4
simplejson==3.17.2
scipy==1.5.4
numpy==1.19.5
hmmlearn==0.2.5
eyeD3==0.9.6
pydub==0.24.1
scikit_learn==0.24.1
tqdm==4.56.0
plotly==4.14.3
python-dotenv==0.15.0
python-magic==0.4.18
